Fix `Disable Mission Weapon Checks` option
Major skill issue from me apparently - I tried getting the ped from the opcode parameters directly instead of getting it from the pool (which is what the original method does) - since I safeguarded the ped pointer in a custom method (`DoesPedHaveWeapon`) it never crashed. In turn, it was never able to properly return whether a ped has a weapon available, resulting in them having NO weapons available when the opcode was called.

Fixes #7